services/api/requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.23.2
python-dotenv==1.0.1
pydantic==2.12.5
pydantic-settings==2.0.2
asyncpg==0.31.0
redis>=5.0.0
elasticsearch==8.11.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0.0,<5.0.0
httpx==0.28.0
email-validator>=2.0.0
python-multipart>=0.0.6
aiohttp>=3.9.0
slowapi>=0.1.9
celery==5.4.0
